% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor_multivariate_quantitative.R
\name{multi.quanti}
\alias{multi.quanti}
\title{Multivarial quantitatives variable analysis to describe clusters}
\usage{
multi.quanti(
  active_variables,
  clusters,
  quali.supp = NULL,
  show_graph = NULL,
  axes = c(1, 2)
)
}
\arguments{
\item{active_variables}{a data frame. Must contains quantitatives variables}

\item{clusters}{a numeric vector. Must contains clusters from a clustering}

\item{quali.supp}{index of supplementary variables in the data frame. Muste be character}

\item{show_graph}{logical. If FALSE, doesn't return any graph}

\item{axes}{choice of dimensions for graph}
}
\value{
List of results (table for individuals and variables, correlation between dimension and clusters)
}
\description{
This function realize multivarial analysis to describe clusters whith numerical variables
}
