% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi_value_test.R
\name{phi.value.test}
\alias{phi.value.test}
\title{Effect size expressed by correlation}
\usage{
phi.value.test(
  active_variables,
  clusters,
  nom = NULL,
  show_graph = TRUE,
  digits = 4
)
}
\arguments{
\item{active_variables}{a factor vector or matrix. Must contains qualitatives variables}

\item{clusters}{a numeric vector. Must contains clusters from a clustering}

\item{nom}{a string. Name of the variable if only one variable is informed in active_variables}

\item{show_graph}{logical. If FALSE, doesn't return any graph}

\item{digits}{a number. Number of digits to display}
}
\value{
The effect size phi for by correlation. Phi is a number between -1 < phi < 1
}
\description{
Returns the effect size phi for by correlation for a modality of interest of a variable and a target group of a cluster
}
\examples{
var = c(rep("yes",7), rep("no",7))
clust = c(1,1,2,1,2,3,1,2,3,3,2,1,3,2)
phi.value.test(var,clust)

}
