% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_multivariate_qualitative.R
\name{plot.multi.quali}
\alias{plot.multi.quali}
\title{Graphics for all the functions of multivariate qualitative analysis}
\usage{
\method{plot}{multi.quali}(x, clusters = FALSE, cramer = FALSE, sup = FALSE, axes = c(1, 2), ...)
}
\arguments{
\item{x}{a data frame. Contains all the data which needs to be print}

\item{clusters}{clusters of the data}

\item{cramer}{v.creamer of the variable with clusters}

\item{sup}{boolean which indicates presence of supplementary variable}

\item{axes}{choice of dimensions}

\item{...}{further arguments of print method}
}
\value{
graphics for the multivariate qualitative analysis
}
\description{
This function allows to print any graphics for the multivariate qualitative analysis. We override the plot function.
}
