% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_multivariate_quantitative.R
\name{plot.multi.quanti}
\alias{plot.multi.quanti}
\title{Graphics for all the functions of multivariate quantitative analysis}
\usage{
\method{plot}{multi.quanti}(x, clusters = NULL, axes = c(1, 2), sup = FALSE, corr = NULL, ...)
}
\arguments{
\item{x}{a data frame. Contains all the data which needs to be print}

\item{clusters}{clusters of the data}

\item{axes}{choice of dimensions}

\item{sup}{boolean which indicates presence of supplementary variable}

\item{corr}{correlation of the variable with clusters}

\item{...}{further arguments of print method}
}
\value{
graphics for the multivariate quantitative analysis
}
\description{
This function allows to print any graphics for the multivariate quantitative analysis. We override the plot function.
}
