% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_univariate_qualitative.R
\name{print.uni.quali}
\alias{print.uni.quali}
\title{Print the results of each function for univariate qualitative}
\usage{
\method{print}{uni.quali}(x, file = NULL, sep = ";", ...)
}
\arguments{
\item{x}{a list}

\item{file}{a string. NULL by default, name for the file to export results}

\item{sep}{a string. ";" by default. Separtor for the file}

\item{...}{parameters passed to other methods}
}
\value{
an array and a file if informed
}
\description{
This functions allows to print all the results of functions for univariate qualitative variables
}
