% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand.R
\name{rand}
\alias{rand}
\title{Compute Rand index between real and predicted clusters}
\usage{
rand(true_label, pred_label)
}
\arguments{
\item{true_label}{a factor vector of real clusters}

\item{pred_label}{a factor vector of predicted clusters}
}
\value{
Rand index
}
\description{
This function returns the Rand index between real clusters and predicted clusters
}
\examples{
true_label <-as.factor(c("1","2","2","1"))
pred_label <-as.factor(c("1","2","1","2"))
rand(true_label,pred_label)

}
