% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cramer_v.R
\name{v.cramer}
\alias{v.cramer}
\title{Compute Cramer's v of qualitatives variables and clusters}
\usage{
v.cramer(active_variables, clusters, show_graph = TRUE, digits = 5)
}
\arguments{
\item{active_variables}{a factor vector or matrix}

\item{clusters}{a numeric vector}

\item{show_graph}{logical. If FALSE, doesn't return graph}

\item{digits}{a number. Number of digits for cramer's v}
}
\value{
The Cramer's v
}
\description{
This function returns the cramer's v from qualitatives variables and clusters
}
\examples{
var = c('yes','no','yes')
clust = c(1,1,2)
v.cramer(var,clust)

}
